Partial LUKS Keyfile Support
The general idea works, but there will be much more work to do in both the library and CLI
before support is fully functional. Namely, we should be able to control the partitioning,
formatting, & mounting of the partition where the keyfile will be stored. The /etc/fstab
file will then need to know where the key partition will be mounted, so that it can be
mounted and searched for.