fix: Use iso-codes package and 'common_name'
The `iso3166-1` crate seems to be outdated and unmaintained. Debian
packages data files with the necessary information, and `distinst`
already depends on various Debian packages, so it's probably better to
use that.

This fixes https://github.com/pop-os/distinst/issues/240 as well as
addressing failures to translate where the string from the `iso3166-1`
crate was not in the gettext translation data.