Activate casefolding support by default
Case-insensitive `ext4` is extremely recent, but the systems this installer's version would be installing are likely to support casefolding.

This feature is useful in cases where a program may not be native to Linux (e.g. being run under WINE [or a derivative]). In addition to activating casefolding itself, enabling case-insensitive reatures requires the user (or automated systems) to add the `CI_dir` attribute to a folder. This *does not* apply system-wide from the get-go, unless the installer was to `chattr +F /` before bootstrapping the system drive.